refactor: utils usage
