+++ src/common/utils/index.ts
@@ -1,8 +1,6 @@
-export const setParam = (paramList: string[][]) => {
-  let params = "";
-  for (let i = 0; i < paramList.length; i++) {
-    if (paramList[i][1] === "") continue;
-    params += `${i === 0 ? "?" : "&"}${paramList[i][0]}=${paramList[i][1]}`;
-  }
-  return params;
-};
+import { type ClassValue, clsx } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
