+++ src/pages/Balance/partials/BalanceBody.tsx
@@ -1,7 +1,7 @@
 import { CircularProgress, Divider } from "@mui/material";
 import { useBalanceContext } from "../context";
 import AppearFadeIn from "@components/Animation/AppearFadeIn";
-import { cn } from "@utils/cn";
+import { cn } from "@utils/index";
 import glassmorphism from "@utils/glassmorphism";
 import BalanceBodyContent from "./BalanceBodyContent";
